iommu/arm-smmu: Fix missing power off for TCU space in testbus handling
When userspace specifies an incorrect testbus sel value, the testbus
write function immediately bails out, leaving the TCU space powered
on, which is not correct. Sanitize the testbus sel value before
programming the TCU testbus registers so that the register space is
powered on only when userspace provides a meaningful value.

Change-Id: I2c0935e056bd411d12ffdcc7d113c628e0e609d4
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>