mink: fix race condition in cb server exit
There could be a race condition between invoke and accept threads of
CB server. If exit for some reason, accept threads would exit and invoke
thread may be processing a CB req that lands up on wait Q after accept
threads have exited. This could cause invoke thread ot hang on wait Q.
Adding a notification mechanism that would set flag if system is either
shutting down or going to reboot. This would help invoke thread to avoid
wait on Q if system was under shutdown when it was about to land on Q.

Change-Id: I8ba942f4ff6284a2ac72e30cec4107ce36e9522f
Signed-off-by: Dinesh K Garg <dineshg@codeaurora.org>