qcacld-3.0: Handle ch cnt and number of channels for SETROAMSCANCHANNELS
To set specific channels, user give SETROAMSCANCHANNELS command with
channel count followed by number of channels.
Ex. SETROAMSCANCHANNELS 2 1 11, here 2 is count and 1 & 11 are the
channels.

If channel count is more than the number of channels, driver rejects
SETROAMSCANCHANNELS command and if number of channels are more than
the channel count, driver ignores excess channels in the list.

Change-Id: I276b84b4994bdf25abf36298dc212dcd54ce0dbb
CRs-Fixed: 2609788