slim-msm-ngd: Check current state of sub system restart notifier
Check for the current state of sub system restart notifier when
the LOCATOR_UP notification is given to slimbus driver and do the
slimbus initialization if the current state is already set to
SERVREG_NOTIF_SERVICE_STATE_UP_V01.

This change is needed in the cases where LOCATOR_UP notification
to slimbus driver got delayed and the state of the sub system restart
notifier is already set to SERVREG_NOTIF_SERVICE_STATE_UP_V01 as
slimbus will be expecting the SERVREG_NOTIF_SERVICE_STATE_UP_V01
notification which is not going to come as the notifier state is
already changed.

Change-Id: I4804bbb13e065ac78be7a442993286532402db9d
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>
Signed-off-by: Chandana Kishori Chiluveru <cchiluve@codeaurora.org>