usb: dwc3: Restart USB if SETUP packet is completed before conndone
In disconnect/connect test scenario, there might be a case when
the SETUP packet from the previous session is completed while the
pullup disable/enable is in progress, i.e., before the Connection
Done event has been fired. This leads to the controller going
into a weird state and not responding to the GET_DEVICE_DESC
request after the Connection Done event. As a result, enumeration
failure is seen.
Fix this by restarting the USB gadget stack when such delayed
SETUP packet is inspected.

Change-Id: I8ef088cb999a7c7472b704da52e602a1e48c67d2
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>