mink: fix incorrect server release
When accept threads are killed, these set server state to DEFUNCT.
If a CB object belonging to now defunct server is release, it would
destroy server while other accept thread might be running. Adding
refcount to server info struct to ensure that server is not destroyed
while others are using it.

Change-Id: Ic53f62badff8bb748238a34247ea9119c40f8844
Signed-off-by: Dinesh K Garg <dineshg@codeaurora.org>