ARM: dts: msm: Enable SMMU stage 1 for Slimbus driver on atoll
This reverts commit 5830190c6df1108bf03c3145ed961e162cd44dd5.
Revert "ARM: dts: msm: Enable SMMU stage 1 for Slimbus driver
on atoll"

This change is reverted as the actual fix is identified and
getting mainlined along with this change. The actual fix is
to check is to add a check for iommu device handler before
doing dma_free_coherent as a part of SSR clean up. This
change ensures we use same dev handle for allocating and
freeing the DMA memory.

Change-Id: If0fa9e1d32e5ccd2a2fee6b0a23930132c536ab1
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>