qcacld-3.0: Pass frequency as an argument to set mon chan api
With recent 6GHz changes set mon chan api has changed to accept
the frequency as an argument instead of the channel number.
Currently existing iwpriv interface is still passing channel
number as an argument to this api which results in the failure
of this command.

To resolve this above issue, convert the channel number received
as an argument to frequency at the entry point of the iwpriv
command and invoke set mon chan api with the fequency as an
argument.

Change-Id: I6bb0682ff3cf5c6e8086012522ea1455a1cecae4
CRs-Fixed: 2675499