i2c-msm-v2: Make init call as module_init for i2c driver
Make init call as module_init for i2c driver instead of
subsys_initcall. This is done to prevent clock off which
happens in i2c probe, as the same clock is shared by uart
(early console). Early console doesn't have a mechanism to
vote for clocks so these clocks should be on till kernel
console driver is probed which is happening as a module_init
call.

Change-Id: I10c6e4f0c5227db97bd5d31dc91c3ba4b9a74038
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>