qcacld-3.0: Validate adapter before netdev features update
Currently a high priority workqueue is scheduled for
enabling the network queue and updating the netdev features
after a successful connection. The adapter validity is not
verified when the workqueue is scheduled, which can lead to
unexpected behaviour.

Validate the adapter in the workqueue before updating
any netdev features.

Change-Id: I29a4dc20993d14c00aef6a4fe502439ae4e14afe
CRs-Fixed: 2776009